requirements.txt
telethon
python-dotenv
solders
solana
requests